# board-system/board-server-tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ResponseModel": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"additionalProperties": true,
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"code"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AuthReqModel": {
				"properties": {
					"clientSecret": {
						"type": "string"
					}
				},
				"required": [
					"clientSecret"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserReqModel": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"nick": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BoardFormModel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"content": {
						"type": "string",
						"nullable": true
					},
					"nick": {
						"type": "string",
						"nullable": true
					},
					"ridx": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"DomainModel": {
				"properties": {
					"host": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"host"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"in": "header",
				"name": "authorization jwt"
			}
		}
	},
	"info": {
		"title": "BoardAPI",
		"version": "1.0.0",
		"description": "TSOA project",
		"license": {
			"name": "Jaehwi-so"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/token": {
			"post": {
				"operationId": "Token_issue",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthReqModel"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login_user",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserReqModel"
							}
						}
					}
				}
			}
		},
		"/board": {
			"get": {
				"operationId": "Select_list",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Insert_board",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BoardFormModel"
							}
						}
					}
				}
			}
		},
		"/board/na": {
			"get": {
				"operationId": "Select_list_na",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Insert_board_na",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BoardFormModel"
							}
						}
					}
				}
			}
		},
		"/board/{id}": {
			"get": {
				"operationId": "Select_one",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update_board",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BoardFormModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete_board",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/board/na/{id}": {
			"get": {
				"operationId": "Select_one_na",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update_board_na",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BoardFormModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete_board_na",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/domain": {
			"post": {
				"operationId": "Insert_domain",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DomainModel"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Select_all",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/resource/single": {
			"post": {
				"operationId": "Insert_resource",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/resource/image": {
			"get": {
				"operationId": "GetImage",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": true,
									"type": "object"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user": {
			"post": {
				"operationId": "Insert_user",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel"
								}
							}
						},
						"description": "Ok"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserReqModel"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8080/api/v1"
		}
	]
}